requirements.txt
numpy-1.14.4
tensorflow==1.5.0
